Code was not working in Frontend and Main
Worked on this to make sure that the css and everything was working right, the only problem is that the button in index.html is being weird. Other than that we fixed everything, next time we merge both branches, I can do it next time as I did it the first time.